Implement InterceptionPatch
InterceptionPatch replaces game bytes with the intention of running a
function.

The implementation uses vector in order to allow for a dynamically sized
buffer to exist until the end of the function's scope. This vector is
used as a buffer to store all the changes that will be made to the
game's bytes.