Replace uintptr_t usage with intptr_t
It is determined that there is no real reason to access memory regions
beyond 2GB. For compatability sake, this region is designated for kernel
space and should not be accessed or modified without permission. If
access is desired, then calling these functions from windows.h is more
appropriate.

This also allows for easier detection of bad addresses.