Move ASM stub to new file and rename D2 function
The ASM stubs are better placed inside of their own file. There are also
macro define checks that allow for architecture agnostic definitions of
stubs.

The D2 function that was added does not appear to have any evidence to
prove that it is a function that exits the game. Therefore, it has been
renamed to Unknown_001 until further notice.