Remove LoD and Classic split for versions < 1.14A
All .dll files are confirmed to be the exact same on Classic and
LoD, when comparing the same version before 1.14A. The only notable
exception is Game.exe, which is different in Classic and LoD, but most
mod writers don't mod that file in memory.

All version from 1.14A and above merged the .dll file into Game.exe, so
their naming schemes are unmodified.

New beta versions have also been added to the list.