Refactor IsGameVersionAtLeast1_14
IsGameVersionAtLeast1_14 now takes a parameter in order to allow
different input.

The function no longer uses unordered set and instead uses a simple
if-statement range check. The check is done on the enum members that
come before 1.14A, so that updating to account for new version is
easier.

The function is also now constexpr for optimization purposes.