Remove Ex function
The Ex function proposal has been declared a bad idea. Functions that
are directly tied to Diablo II functions do not need unnecessary
abstractions that fail even at that purpose. They should be at the right
level so that they are abstract enough for high level usage, but
practical and usable for low level assembly usage.