Update the license to be more specific
The text now indicates permitted linking with the game and its
libraries, rather than specifying it as just a library.

Diablo II has been clrified to also include Lord of Destruction. In the
case that any new expansions or remake are released where it shares most
of the same code, a license change will be done for clarification. This
will be done by requiring that external committers be required to sign
a written agreement to permit expansion of the license for this purpose.