Refactor Game Version, add C interface
The singleton approach was not a good idea as originally thought. More
code has to be written to express the same information, especially for
code that depends on the running game version.

Instead of forcing developers to deal with the annoyance of dealing with
the singleton and the verbosity, the singleton is now in an unnamed
namespace, where multiple public functions make access to the singleton
easier and intuitive.