Revise C++ API to new specifications
The original game struct is defined with the original name and acts as
the base for the version specific implementations.

The Wrapper simply allows use of the structs with getters. No other
functionality is provided.

The API abstraction class allows for the creation of instances of the
struct and provides additional features for object-oriented usage.