Statically allocate file version table
This uses a similar solution to the C89 repo, but written in modern C++.
This uses std::array instead of std::unordered_map for lookup.

It may be slower because O(1) is slower than O(log(n)). However, the
allocation is constant, so no unnecessary initialization is required. In
addition, the slowdown is negligent since the table element count never
increases during runtime, so time complexity has an actual constant
expected worst time. Also, the lookup is only done once, so the overall
impact is minimal.